// generated/src/app/api-models/cap-trans.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapTrans {
  seqCapTrans: number;
  capModelId: string;
  firstContractMonth: Date;
  capEntityCode: string;
  seqCapPoolId: number;
  capMonthRun: Date;
  capMonth: Date;
  seqProvId: number;
  seqVendId: number;
  transType: string;
  capStoplossId: string;
  seqCapAdjustment: number;
  processingStatus: string;
  capTransAmt: number;
  pmpmWitholdAmt: number;
  postingRule: string;
  seqApTrans: number;
  postDate: Date;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  checkDate: Date;
  seqCapVendAddress: number;
  adjustmentFlag: string;
  accessIncentiveIncluded: string;
  qualityIncentiveIncluded: string;
  fileType: string;
}